+++ library/src/main/java/org/mustangproject/Item.java
@@ -176,15 +176,25 @@ public class Item implements IZUGFeRDExportableItem {
 				stac.getAsNodeMap("ChargeIndicator").ifPresent(ci -> {
 					String isChargeString=ci.getAsString("Indicator").get();
 					String percentString=stac.getAsStringOrNull("CalculationPercent");
+					String amountString=stac.getAsStringOrNull("ActualAmount");
 					String reason=stac.getAsStringOrNull("Reason");
-					Allowance izac=null;
+					Charge izac= new Charge();
 					if (isChargeString.equalsIgnoreCase("false")) {
 						izac = new Allowance();
+					} else {
+						izac = new Charge();
 					}
-
+					if (amountString!=null) {
+						izac.setTotalAmount(new BigDecimal(amountString));
+					}
 					izac.setPercent(new BigDecimal(percentString));
 					izac.setReason(reason);
-					addAllowance(izac);
+
+					if (isChargeString.equalsIgnoreCase("false")) {
+						addAllowance(izac);
+					} else {
+						addCharge(izac);
+					}
 				});
 
 			});