+++ library/src/test/java/org/mustangproject/ZUGFeRD/ZF2ZInvoiceImporterTest.java
@@ -40,7 +40,6 @@ import java.nio.file.Paths;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Arrays;
-import java.util.Date;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -102,31 +101,6 @@ public class ZF2ZInvoiceImporterTest extends ResourceCase {
 		// name street location zip country, contact name phone email, total amount
 
 	}
-
-
-	public void testTheInvoiceImport() {
-
-		ZUGFeRDInvoiceImporter zii = new ZUGFeRDInvoiceImporter("C:\\Users\\jstaerk\\workspace\\XMLExamples\\zfdiverses\\20250108_\\xmlDieWirLesenSollten.xml");
-
-		boolean hasExceptions = false;
-		Invoice invoice = null;
-		try {
-			invoice = zii.extractInvoice();
-		} catch (XPathExpressionException | ParseException e) {
-			hasExceptions = true;
-		}
-		assertFalse(hasExceptions);
-		// Reading ZUGFeRD
-
-		TransactionCalculator tc = new TransactionCalculator(invoice);
-		assertEquals(new BigDecimal("0.00"), tc.getDuePayable());
-		assertEquals(new BigDecimal("0.00"), tc.getGrandTotal());
-
-
-		// name street location zip country, contact name phone email, total amount
-
-	}
-
 
 
 	public void testInvoiceImportUBL() {