+++ History.md
@@ -8,7 +8,8 @@ allow to add includedNotes with type
 - 631  multiple invoice referenced documents
 - 629
 - 630 #296 #565
-- 
+- 657
+- 658
 
 2.15.2
 =======