+++ History.md
@@ -8,7 +8,14 @@ allow to add includedNotes with type
 - 631  multiple invoice referenced documents
 - 629
 - 630 #296 #565
-- 
+- 657
+- 658
+- 648
+- 651
+- 652
+- -620
+- 653
+- 654
 
 2.15.2
 =======